+++ src/test/clojurescript/clara/test.cljs
@@ -19,7 +19,10 @@
             [clara.test-simple-rules]
             [clara.test-rhs-retract]
             [clara.test-bindings]
-            [clara.test-clear-ns-productions]))
+            [clara.test-clear-ns-productions]
+            [clara.test-negation]
+            [clara.performance.test-rule-execution]
+            [clara.test-node-sharing]))
 
 (enable-console-print!)
 
@@ -54,5 +57,8 @@
                     'clara.test-simple-rules
                     'clara.test-rhs-retract
                     'clara.test-bindings
-                    'clara.test-clear-ns-productions)
+                    'clara.test-clear-ns-productions
+                    'clara.test-negation
+                    'clara.performance.test-rule-execution
+                    'clara.test-node-sharing)
     @*successful?*))